\name{check_temperature}
\alias{check_temperature}
\title{
Function to check the consistency of the temperature data stored in haul data table (TA).}
\description{
This function checks if the temperature by haul is in the range 10-30 Celsius degrees; moreover, a plot depth versus temperature is produced and stored in the Graph folder.}
\usage{
check_temperature(ResultDataTA,year,wd,suffix)
}
\arguments{
\item{ResultDataTA}{Haul data table according to MEDITS protocol (TA).}
\item{year}{reference year for the analysis.}
\item{wd}{Working directory selected by the user.}
\item{suffix}{Suffix string of the Logfile.}
}
\details{
This check uses the temperature range 10-30 Celsius degrees to provide quantitative warning and a plot, automatically stored in Graphs, for a qualitative inspection  of that temperature data respect to depth. }
\value{
The function returns always TRUE, because the outcome of the function is a warning that does not lock the execution of the 'RoME' checks. The presence of inconsistencies in the data is reported in the logfile stored in the "Logfiles" subdirectory of the "wd" user-defined directory.
}
\references{
Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp. \url{https://www.sibm.it/MEDITS\%202011/principaledownload.htm}
}
\author{
I. Bitetto, W. Zupa
}
\examples{
TA = RoME::TA
year=2012
wd=tempdir()
suffix="2020-03-05_time_h17m44s55"
check_temperature(TA,year,wd,suffix)
}
\keyword{temperature}


