\name{GSAs}
\alias{GSAs}
\docType{data}
\title{List of GFCM Geographical subareas (GSAs)}
\description{GSAs table}
\usage{data("GSAs")}
\format{
  A data frame with 31 observations on the following 3 variables.
  \describe{
    \item{\code{GSA}}{a numeric vector}
    \item{\code{CODE}}{a character vector}
    \item{\code{Area}}{a character vector}
  }
}

\source{
\url{https://www.fao.org/gfcm/data/maps/gsas/en/}
}
\references{
\url{https://www.fao.org/gfcm/data/maps/gsas/en/}}
\examples{
data(GSAs)
str(GSAs)
head(GSAs)
}
\keyword{datasets}
