% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.conversion.r
\name{headers.conversion}
\alias{headers.conversion}
\title{Headers conversion for MEDITS tables}
\usage{
headers.conversion(table, type, verbose = FALSE)
}
\arguments{
\item{table}{data frame of the TX table}

\item{type}{type of tables to be analysed. Allowed values: "TA","TB","TC","TE","TL"}

\item{verbose}{boolean. If TRUE a message is printed.}
}
\value{
A data frame object is returned including the only allowed field
}
\description{
Headers conversion for MEDITS tables
}
\details{
The functions allow to convert headers of table coming from RDBFIS data base to the MEDITS format expected from RoME package
}
