\name{list_g1_g2}
\alias{list_g1_g2}
\docType{data}
\title{List of G1 and G2 species}
\description{List of the target species G1 and G1 as defined by the MEDITS protocol (see MEDITS-Handbook Version 9 2017)}
\usage{data("list_g1_g2")}
\format{
  A data frame with 88 observations on the following 17 variables.
  \describe{
    \item{\code{No}}{a numeric vector of progressive number}
    \item{\code{Medit_LIST_proposal_2011}}{a factor with levels of the list proposed in 2011}
    \item{\code{Species_group_DCF}}{a factor with levels of the DCF species groups}
    \item{\code{MEDITS_G1}}{a numeric vector of G1 species}
    \item{\code{MEDITS_G2}}{a numeric vector of G2 species}
    \item{\code{Group}}{a factor with levels of groups}
    \item{\code{Old_MEDITS_list}}{a numeric vectorof the old MEDITS list}
    \item{\code{Tot_No}}{Tot_No}
    \item{\code{Tot_W}}{Tot_W}
    \item{\code{Ind_Length}}{Ind_Length}
    \item{\code{Sex}}{a factor with levels of sex}
    \item{\code{Mat_stage}}{a factor with levels of maturity stages}
    \item{\code{Age}}{a factor with levels of age}
    \item{\code{Ind_weight}}{a factor with levels of Ind_weight}
    \item{\code{Date}}{a factor with levels of Date}
    \item{\code{CODE}}{a factor with levels CODE}
    \item{\code{English_common_name}}{a factor with levels of common names in english language}
  }
}

\source{Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp.\url{https://www.sibm.it/MEDITS\%202011/principaledownload.htm}}
\references{Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp.\url{https://www.sibm.it/MEDITS\%202011/principaledownload.htm}}
\examples{
data(list_g1_g2)
}
\author{W. Zupa}
\keyword{datasets}


