\name{RoMEcc}
\alias{RoMEcc}
\title{Continuous Quality Checks on MEDITS Trawl Survey Data}
\description{
The function calls all the functions included in the package in an ordered way to perform a complete quality check of TX data available. The check is performed simultaneously on the files that can contain also data of more than one year. Unlike the \code{RoME} function, \code{RoMEcc} does not stop at the first detected error, allowing user to correct data. Instead, it checks all the data and returns a report on the errors found, as well as compiling a detailed log file.}
\usage{
RoMEcc(TA,TB,TC,TE=NA,TL=NA,wd,suffix=NA,
verbose =TRUE,Stratification=RoME::stratification_scheme,
Ref_list=RoME::TM_list,DataTargetSpecies=RoME::DataTargetSpecies,
Maturity=RoME::Maturity_parameters,
ab_parameters=RoME::LW,
stages_list=RoME::mat_stages,assTL=RoME::assTL, zip=TRUE)
}
\arguments{
  \item{TA}{Haul data table according to MEDITS protocol (TA)}
  \item{TB}{Catch data table according to MEDITS protocol (TB)}
  \item{TC}{Biological data table according to MEDITS protocol (TC)}
  \item{TE}{Individual biological data table according to MEDITS protocol (TE) if available, if TE data are not available, use NA.}
  \item{TL}{Litter data table according to MEDITS protocol (TL) if available, if TL data are not available, use NA}
  \item{wd}{working directory path defined by the user}
  \item{suffix}{Suffix string of the Logfile, in case it \code{suffix=NA} it is automatically generated by the function using date and time of the check}
  \item{verbose}{...}
  \item{Stratification}{Stratification scheme according to MEDITS protocol.}
  \item{Ref_list}{TM_list reference list}
  \item{DataTargetSpecies}{Information related to target species.}
  \item{Maturity}{Information related to sex and maturity from literature or other sources.}
  \item{ab_parameters}{dataframe containing the a and b parameters of the length-weight relationships}
  \item{stages_list}{Table of maturity stages.}
  \item{assTL}{data frame with the association between TL (litter table) categories and sub-categories}
  \item{zip}{boolean, if TRUE a zip file containing the results is generated}
}
\details{
\code{RoMEcc} checks can be used to integrate a list of common quality checks on survey data. This function calls all the functions built in the package in an ordered way to perform a complete quality check of TX data available. The order of the checks in \code{RoME} was implemented in a defined sequence to avoid cascade errors due to the correction of a previous error. No automatic correction is implemented in 'RoME'. 'RoME' stops if an error occurs; then the user has to correct the error and run again the code to continue with the other checks. The function runs on a complete time series dataset, checking year after year, until the end of the time series. After the checks of the mandatory fields and the controlled vocabulary, that are carried out for all the TX tables, the specific checks on each kind of TX table are performed. Finally, \code{RoME} provides a list of cross checks aimed to guarantee the consistency among the data tables.

Some functions included in the 'RoME' library and used by \code{RoME} function need specific dictionaries or tables. It is the case of \code{Stratification}, \code{TM_list}, \code{DataTargetSpecies}, \code{Maturity_parameters}, \code{mat_stages} and \code{assTL} tables. All of them are provided by default in this library. Anyway, the user has the possibility to provide \emph{ad hoc} modified versions of these tables adapting the checks to specific needs.
}
\value{
The function does not correct data, but it detects the errors, warning the user that there is the possibility of one or more errors, specifying the type of the error and easing the data correction. If parameter \code{verbose=TRUE} returns a series of text output in console to let the user to trace the state of the checks. All the output of the functions are stored in the user defined working directory  \code{wd} and in the sub-directory there resident. In the Lofile subfolder are stored the logfiles of each run of the function.
}
\references{Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp. \url{https://www.sibm.it/MEDITS\%202011/principaledownload.htm}}
\author{W. Zupa,I. Bitetto}
\examples{
\donttest{
wd <- tempdir()
suffix=NA
DataTA = data.frame(RoME::TA[RoME::TA$YEAR==2012 ,])
DataTB = data.frame(RoME::TB[RoME::TB$YEAR==2012 ,])
DataTC = data.frame(RoME::TC[RoME::TC$YEAR==2012 ,])
DataTE = NA
DataTL = NA
RoMEcc(DataTA, DataTB,DataTC,DataTE,DataTL, wd, suffix,
  Stratification=RoME::stratification_scheme,
  Ref_list=RoME::TM_list,DataTargetSpecies=RoME::DataTargetSpecies,
  Maturity=RoME::Maturity_parameters,ab_parameters=RoME::LW,
  stages_list=RoME::mat_stages,assTL=RoME::assTL, verbose=FALSE)
}
}

\keyword{error}


