\name{templateTL}
\alias{templateTL}
\docType{data}
\title{
TL table template
}
\description{
Template of the Litter data table (TL) as defined in the MEDITS protocol
}
\usage{data("templateTL")}
\format{
  A data frame with 0 observations on the following 14 variables.
  \describe{
    \item{\code{TYPE_OF_FILE}}{a logical vector}
    \item{\code{COUNTRY}}{a logical vector}
    \item{\code{AREA}}{a logical vector}
    \item{\code{VESSEL}}{a logical vector}
    \item{\code{YEAR}}{a logical vector}
    \item{\code{MONTH}}{a logical vector}
    \item{\code{DAY}}{a logical vector}
    \item{\code{HAUL_NUMBER}}{a logical vector}
    \item{\code{LITTER_CATEGORY}}{a logical vector}
    \item{\samp{LITTER_SUB-CATEGORY}}{a logical vector}
    \item{\code{TOTAL_WEIGHT_IN_THE_CATEGORY_HAUL}}{a logical vector}
    \item{\code{TOTAL_NUMBER_IN_THE_CATEGORY_HAUL}}{a logical vector}
    \item{\samp{TOTAL_WEIGHT_IN_THE_SUB-CATEGORY_HAUL}}{a logical vector}
    \item{\samp{TOTAL_NUMBER_IN_THE_SUB-CATEGORY_HAUL}}{a logical vector}
  }
}
\details{For details see MEDITS Survey - Instruction Manual - Version 9 (2017)}
\source{MEDITS Survey - Instruction Manual - Version 9 (2017)}
\references{Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp.\url{https://www.sibm.it/MEDITS\%202011/principaledownload.htm}}
\examples{
\donttest{data(templateTL)}
}
\author{W. Zupa}
\keyword{datasets}


