\name{time}
\alias{time}
\docType{data}
\title{
allowed values for SHOOTING_TIME and HAULING_TIME
}
\description{
allowed values for SHOOTING_TIME and HAULING_TIME
}
\usage{data("time")}
\format{
  A data frame with allowed values for SHOOTING_TIME and HAULING_TIME.
}
\details{
The integer values vector is used to check the correctness of SHOOTING_TIME and HAULING_TIME.
}
\source{
Anonymus (2017) "MEDITS-Handbook. Version n. 9. MEDITS Working Group" \url{https://www.sibm.it/MEDITS\%202011/principaledownload.htm}
}
\references{
Anonymus (2017) "MEDITS-Handbook. Version n. 9. MEDITS Working Group" \url{https://www.sibm.it/MEDITS\%202011/principaledownload.htm}
}
\examples{
data(time)
str(time)
}
\author{W. Zupa}
\keyword{datasets}



