\name{classes}
\alias{classes}
\docType{data}
\title{Class of fields}
\description{Definition of field' classes for TX tables}
\usage{data("classes")}
\format{
  A data frame with 123 observations on the following 4 variables.
  \describe{
    \item{\code{RDBFIS}}{a character vector}
    \item{\code{MEDITS}}{a character vector}
    \item{\code{table}}{a character vector}
    \item{\code{type}}{a character vector}
  }
}
\details{
See Medits handbook.
}
\references{
Anonymus. 2017. MEDITS-Handbook. Version n. 9. MEDITS Working Group. 106 pp. sibm.it/MEDITS 2011/principaledownload.htm
}
\examples{
\donttest{data(classes)}
}
\keyword{datasets}
